Rename stateTracker.ts
