+++ packages/ag-charts-community/src/util/stateTracker.ts
